feat(ui): implement create identifier
