+++ src/ui/components/CreateIdentifier/components/IdentifierThemeSelector/IdentifierThemeSelector.types.ts
@@ -1,0 +1,14 @@
+import { IdentifierColor } from "../IdentifierColorSelector";
+
+interface ThemeItemProps {
+    index: number;
+    color: number;
+}
+
+interface IdentifierThemeSelectorProps {
+  color: IdentifierColor;
+  selectedTheme: number;
+  setSelectedTheme: (value: number) => void;
+}
+
+export type { ThemeItemProps, IdentifierThemeSelectorProps }
