+++ src/ui/components/CreateIdentifier/components/IdentifierThemeSelector/index.ts
@@ -1,0 +1,1 @@
+export * from "./IdentifierThemeSelector";
