+++ src/ui/components/CreateIdentifier/components/TypeItem/TypeItem.types.ts
@@ -1,0 +1,10 @@
+interface TypeItemProps {
+    dataTestId: string;
+    index: number;
+    text: string;
+    clickEvent: (index: number) => void;
+    selectedType: number;
+    disabled?: boolean;
+}
+
+export type { TypeItemProps };
